percpu: fix printf warning about int vs long
git-svn-id: svn://localhost/svn/linux-kernel/trunk@9758 526b6c2d-f592-4532-a319-5dd88ccb003d

+++ mm/percpu.c
@@ -1647,7 +1647,7 @@ int __init pcpu_embed_first_chunk(size_t reserved_size, size_t dyn_size,
 	if (max_distance > (VMALLOC_END - VMALLOC_START) * 3 / 4) {
 		pr_warning("PERCPU: max_distance=0x%zx too large for vmalloc "
 			   "space 0x%lx\n",
-			   max_distance, VMALLOC_END - VMALLOC_START);
+			   max_distance, (unsigned long)(VMALLOC_END - VMALLOC_START));
 #ifdef CONFIG_NEED_PER_CPU_PAGE_FIRST_CHUNK
 		/* and fail if we have fallback */
 		rc = -EINVAL;
